lock aws-sdk version and expose deployment functions via require
